Created migrations. Migrations see inside ApplicationDbContext and see that we have a DbSet on Category and will write a script to generate SQL to create the table itself. Installed EFC.Tools. In order to run migrations we need run NuGet PM via console and run "add-migration AddCategoryToDb". It will create a new folder Migrations and load the migrations.
Added comma that was preventing execution of add-migration in PMC.

Once migration is created and loaded, review it to make sure everything looks good according to what we set in the model and then...

Ran "update-database" in PMC. This will first check via connection string if the db exists. If it does then it will push the migration we just added; if it doesn't it will create the db then it'll run the migration. Since we didn't already have a db, running this update will create one for us and then create the tables.

EFMigrationHistory is also created and updated and will show all migration Ids (name of the migration in VS) and product version. This way EFC will know that the migration has already been applied, so next time you run the command "update-database" it will not reapply the migration. This is a way we can track everything done with EF.